New slopes object
