+++ arte/types/slopes.py
@@ -2,44 +2,101 @@ import numpy as np
 
 
 class Slopes(object):
+    ''''
+    Class that represents one or more generic WFS slope frames
+    '''
+    def __init__(self, slopesx_compressed, slopesy_compressed, mask2d):
+        '''
+        Parameters
+        ----------
+        slopesx_compressed: np.ndarray
+            1D or 2D slope vector. If 2D, it is assumed to be [time, slopes]
+        slopesy_compressed: np.ndarray
+            1D or 2D slope vector. If 2D, it is assumed to be [time, slopes]
+        mask: np.ndarray
+            2D boolean mask, valid for both X and Y slope vectors
+        '''
+        self._mask2d = mask2d
+        self._vector = np.hstack((slopesx_compressed, slopesy_compressed))
+        if len(self._vector.shape) == 1:
+            self._vector = np.expand_dims(self._vector, axis=0)
 
-    def __init__(self, mapX, mapY):
+    @staticmethod
+    def from2dmaps(mapX, mapY):
+        '''
+        Build a Slopes object from X/Y 2d maps or a series of X/Y 2d maps.
+        Both maps must be np.ma.MaskedArray objects. Masks for X and Y slopes
+        must be identical.
+
+        Parameters
+        ----------
+        mapX: np.ma.MaskedArray
+            2D or 3D slope map. If 3D, it is assumed to be [time, rows, cols]
+        mapY: np.ma.MaskedArray
+            2D or 3D slope map. If 3D, it is assumed to be [time, rows, cols]
+
+        Returns
+        -------
+        slopes: Slopes object
+            A new Slopes object. Valid slopes (as indicated by the mask)
+            is copied into an internal array. No references to the input data are kept.
+        '''
         assert isinstance(mapX, np.ma.MaskedArray)
         assert isinstance(mapY, np.ma.MaskedArray)
-        self._mapX = mapX
-        self._mapY = mapY
-        self._vector = np.hstack((self._mapX.compressed(),
-                                 self._mapY.compressed()))
-        self._nOfSlopes = self._vector.size
+        assert mapX.shape == mapY.shape
+        assert np.all(mapX.mask == mapY.mask)
+
+        mask = mapX.mask.copy()
+        if len(mapX.shape) == 2:
+            mask = np.expand_dims(mask, axis=0)
+
+        nframes = mapX.shape[0]
+        mask2d = np.logical_or.reduce(mask, axis=0)
+
+        nOfSubaps = np.count_nonzero(~mask2d)
+        return Slopes(mapX.compressed().reshape(nframes, nOfSubaps),
+                      mapY.compressed().reshape(nframes, nOfSubaps),
+                      mask2d)
 
     @staticmethod
     def fromNumpyArray(mapXAsMaskedNumpyArray,
                        mapYAsMaskedNumpyArray):
-        return Slopes(mapXAsMaskedNumpyArray,
-                      mapYAsMaskedNumpyArray)
+        return Slopes.from2dmaps(mapXAsMaskedNumpyArray,
+                                 mapYAsMaskedNumpyArray)
 
     def toNumpyArray(self):
-        return self._mapX, self._mapY
+        return self.mapX(), self.mapY()
 
-#    def pupils(self):
-#        return self._pupils
     def numberOfSlopes(self):
-        return self._nOfSlopes
+        return self._vector.shape[1]
+
+    def numberOfFrames(self):
+        return len(self._vector)
+
+    @property
+    def shape2d(self):
+        return self._mask2d.shape
 
     def mapX(self):
-        return self._mapX
+        fullshape = (self.numberOfFrames(),) + self.shape2d
+        data = np.zeros_like(self._vector, shape=fullshape)
+        data[:, *np.where(~self._mask2d)] = self._vector[:, :self.numberOfSlopes()//2]
+        return np.squeeze(np.ma.array(data, mask=np.broadcast_to(self._mask2d, shape=fullshape)))
 
     def mapY(self):
-        return self._mapY
+        fullshape = (self.numberOfFrames(),) + self.shape2d
+        data = np.zeros_like(self._vector, shape=fullshape)
+        data[:, *np.where(~self._mask2d)] = self._vector[:, self.numberOfSlopes()//2:]
+        return np.squeeze(np.ma.array(data, mask=np.broadcast_to(self._mask2d, shape=fullshape)))
 
     def vector(self):
-        return self._vector
+        return np.squeeze(self._vector)
 
     def vectorX(self):
-        return self._vector[0: self._nOfSlopes // 2]
+        return np.squeeze(self._vector[:, 0: self.numberOfSlopes() // 2])
 
     def vectorY(self):
-        return self._vector[self._nOfSlopes // 2:]
+        return np.squeeze(self._vector[:, self.numberOfSlopes() // 2:])
 
     def __eq__(self, o):
         if not isinstance(o, Slopes):
